+++ njunmt/encoders/transformer_encoder.py
@@ -51,14 +51,15 @@ class TransformerEncoder(Encoder):
             self._self_attention_layers.append(
                 MultiHeadAttention(self.params["selfattention.params"], self.mode))
 
-        if self.mode == ModeKeys.INFER:
-            self.encoder_output_tuple_type = namedtuple(
-                "EncoderOutput",
-                "outputs attention_values attention_length encoder_self_attention")
-        else:
+        if self.mode == ModeKeys.TRAIN:
             self.encoder_output_tuple_type = namedtuple(
                 "EncoderOutput",
                 "outputs attention_values attention_length")
+        else:
+            self.encoder_output_tuple_type = namedtuple(
+                "EncoderOutput",
+                "outputs attention_values attention_length encoder_self_attention")
+
 
     @staticmethod
     def default_params():
@@ -91,19 +92,20 @@ class TransformerEncoder(Encoder):
             # [batch_size, 1, 1, timesteps], FLOAT_MIN for padding, 0.0 for non-padding
             encoder_attention_bias = MultiHeadAttention.attention_length_to_bias(inputs, feature_length)
             outputs, enc_self_attention = self._transform(inputs, encoder_attention_bias, scope=vs, **kwargs)
-            if self.mode == ModeKeys.INFER:
-                encoder_output = self.encoder_output_tuple_type(
-                    # [batch_size, timesteps, dim]
-                    outputs=outputs,
-                    attention_values=outputs,
-                    attention_length=feature_length,
-                    encoder_self_attention=enc_self_attention)
-            else:
+            if self.mode == ModeKeys.TRAIN:
                 encoder_output = self.encoder_output_tuple_type(
                     # [batch_size, timesteps, dim]
                     outputs=outputs,
                     attention_values=outputs,
                     attention_length=feature_length)
+            else:
+                encoder_output = self.encoder_output_tuple_type(
+                    # [batch_size, timesteps, dim]
+                    outputs=outputs,
+                    attention_values=outputs,
+                    attention_length=feature_length,
+                    # a list of Tensors, [batch_size, num_heads, length_q, length_k]
+                    encoder_self_attention=enc_self_attention)
             return encoder_output
 
     def _transform(self, inputs, encoder_self_attention_bias, **kwargs):
@@ -122,7 +124,7 @@ class TransformerEncoder(Encoder):
         input_padding = attention_bias_to_padding(encoder_self_attention_bias)
         pad_remover = PadRemover(input_padding)
         x = dropout_wrapper(inputs, self.params["layer_prepostprocess_dropout_keep_prob"])
-        cache_atts = []
+        encoder_self_attention_scores = []
         for layer in range(self.params["num_layers"]):
             with tf.variable_scope("layer_%d" % layer):
                 with tf.variable_scope("self_attention"):
@@ -133,7 +135,7 @@ class TransformerEncoder(Encoder):
                             x=x, process_sequence=self.params["layer_preprocess_sequence"],
                             dropout_keep_prob=self.params["layer_prepostprocess_dropout_keep_prob"]),
                         memory_bias=encoder_self_attention_bias)
-                    cache_atts.append(w_y)
+                    encoder_self_attention_scores.append(w_y)
                     # apply dropout, layer norm, residual
                     x = layer_postprocessing(
                         x=y, previous_x=x,
@@ -156,4 +158,4 @@ class TransformerEncoder(Encoder):
         x = layer_preprocess(
             x=x, process_sequence=self.params["layer_preprocess_sequence"],
             dropout_keep_prob=self.params["layer_prepostprocess_dropout_keep_prob"])
-        return x, cache_atts
+        return x, encoder_self_attention_scores